Rewritten to be an API and called as such.
This is the final prerelease version of the item frames for MineClone 2.
This version has fully working frames and glow frames. It is backwards compatible with all of the previous versions of the item frames. 
----------
Added in backwards compatibility. Made the basic frame definitions into an API. Made 2 frames that call the API - The glow frame and the standard frame.
Wrote a rotate displayed item function to add in the missing functionality. Item Frames now require screwdrivers to rotate the displayed items, but it does not cause wear on the screwdriver, as it was meant to be inherent functionality.  (Required the screwdriver's On_Rotate callback to create the functionality.)

Tested on a single server. After full review, the approved code and translations will be put up.